move Settings into subfolder of /views
